Letter Manager setup  + GameManagerScorefunction + OverallUiChanges
Now the game has the letter controller on the Hud object, that's where the letters of the level are instantiated, also added a feedback for the word typed by the player. Added the score function to give score to the player. Made some adjustments on the folders setup and namespaces.