Save Data
Implemented the save manager which handles the saving using binary serialization. The journey score and the last completed level are data saved in the file. The Journey score is the cumulative score from all completed levels, which is the saved one. The level also has a Score. which is valid only for each level, in case the player exit without completing the crossword the level score is not added to the journey score.