Removing Level selection Options
I've added this feature thinking on a scenario where player could replay the same level. The idea was to let them search for bonus words if they want to, but it would increase too much of the scope with no much time to work on this would have to save every success attempt of word player did in each level. And overall, it seams ok without it.